Add proof of concept for CSS fragments
Using CSS fragments, one can add or remove CSS classes using the same
mechanism as when adding step-by-step content. In addition to the
"fragment" class, one needs to add a "css-fragment" class as well as a
"data-target" attribute which specifies the target for the css class
addition or removal.
The latter are specified using "data-add-class" and "data-remove-class".