Updated console instructions for launching notes
When running the notes-server plugin, the node console had a message
that said to look in the JS console for a link to the notes page.  The
JS console log message no longer appears, so I have updated the node
console message to reflect how the notes page should be opened (pressing
the 's' key) as described in the documentation.