Fontawesome - fixes
As Fontawesome uses the <i>-Tag, I had to remove reset of <i> (line 21 - this was because of fontawesome font sizes) as well as remove the tag for the italic definition (line 251). Now you can easily add Fontawesome to your slides by putting <link rel="stylesheet" href="css/font-awesome.css"> into the head of the html-file.