Suporting unicode id's
Don't care about the actual content of id. HTML4 cares, but html5 does not. Even, in html4, is pointless manufacturing a new id to target that indeed it is not there. If the id is not legal searching for it will fail and the effect would be the current behavior.

It fixes  #2198. About pandoc generated ids based on unicode titles.