Limit index in updateFragments()
This avoids getting in a state where the fragment index is not -1 but no fragments are marked as current. This can happen when nagivating down from the last fragment at the bottom of a vertical stack of slides or from an outdated URL hash. This is most noticeable with a highlight-current fragment style on the last fragment since it will become unhighlighted.