This commit allows the user to pass an optional filename argument
It is now possible to pass an optional filename argument for specifying
the name of the presentation html file. If no argument is passed, the
default behavior is the same, i.e. open index.html.
If started with npm start -- --file=presentation.html then the file
presentation.html is served. This file might be located in a different
directory. Makes it easy to store separate presentations in another directory
or even repository.